feat(component): WIP to make ExperimentDetails component smaller
